+++ lib/crypto/ecdsa.js
@@ -2,16 +2,18 @@
 
 var BN = require('./bn');
 var Point = require('./point');
+var Signature = require('./signature');
 var Random = require('./random');
 var PublicKey = require('../publickey');
 var PrivateKey = require('../privatekey');
-var Signature = require('../signature');
 
 var ECDSA = function ECDSA(obj) {
-  if (!(this instanceof ECDSA))
+  if (!(this instanceof ECDSA)) {
     return new ECDSA(obj);
-  if (obj)
+  }
+  if (obj) {
     this.set(obj);
+  }
 };
 
 ECDSA.prototype.set = function(obj) {
