+++ lib/crypto/signature.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var BN = require('./crypto/bn');
+var BN = require('./bn');
 
 var Signature = function Signature(r, s) {
   if (!(this instanceof Signature))
